task

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -2,11 +2,11 @@
 /**
  * print_listint - function that prints all the elements of a listint_t list
  * @h : it's a pointer that pointer on the head of the list
- * Return: return the elements of the list
+ * Return: list
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t counter;
+	int counter;
 	if (h == NULL)
 	{
